requirements.txt
#apache-airflow==2.2.3
numpy==1.19.5
pandas==1.1.5
#psycopg2==2.9.7
psycopg2-binary==2.9.2
SQLAlchemy==1.4.54
textblob==0.17.1
